delays and frequencies types rework

+++ lib_arch_xt/src/arch_xt_adc.h
@@ -56,9 +56,9 @@ struct ArchXT_ADCConfig {
     std::vector<ADC::channel_config_t> channels;
     std::vector<reference_config_t> references;
     uint32_t vref_channel;
-    std::vector<uint16_t> clk_ps_factors;
-    uint16_t clk_ps_max;
-    std::vector<uint16_t> init_delays;
+    std::vector<unsigned long> clk_ps_factors;
+    unsigned long clk_ps_max;
+    std::vector<unsigned long> init_delays;
 
     reg_addr_t reg_base;
